entidades arrumadas com novas relações e dados

+++ src/main/java/com/unip/safeEats/API/repositories/FormaPagamentoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormaPagamentoRepository extends JpaRepository<FormaPagamento, Long> {
+public interface FormaPagamentoRepository extends JpaRepository<FormaPagamento, Integer> {
 }
