+++ src/main/java/com/unip/safeEats/API/repositories/UsuarioRepository.java
@@ -4,6 +4,8 @@ import com.unip.safeEats.API.entities.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 }
